requirements.txt
gspread
google-auth
google-auth-oauthlib
pandas
playwright
